5_parallel_execution/fn_create_cloud_tasks/requirements.txt
-i https://pypi.org/simple
cachetools==5.2.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
cloudevents==1.8.0
deprecation==2.1.0
flask==2.2.2 ; python_version >= '3.7'
functions-framework==3.3.0
google-api-core[grpc]==2.11.0 ; python_version >= '3.7'
google-auth==2.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-tasks==2.11.0
googleapis-common-protos==1.57.0 ; python_version >= '3.7'
grpc-google-iam-v1==0.12.4 ; python_version >= '3.6'
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0 ; platform_system != 'Windows'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==5.2.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
packaging==22.0 ; python_version >= '3.7'
proto-plus==1.22.1 ; python_version >= '3.6'
protobuf==4.21.12 ; python_version >= '3.7'
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
rsa==4.9 ; python_version >= '3.6'
setuptools==65.6.3 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
types-click==7.1.8
types-flask==1.1.6
types-jinja2==2.11.9
types-markupsafe==1.1.10
types-werkzeug==1.0.9
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
watchdog==2.2.0 ; python_version >= '3.6'
werkzeug==2.2.2 ; python_version >= '3.7'
zipp==3.11.0 ; python_version >= '3.7'
